requirements.txt
moviepy==1.0.3
Pillow==10.4.0
numpy==2.0.1
imageio==2.35.1
imageio-ffmpeg==0.5.1

